feat: Cleanup
